Modified SkinningMesh implementation for efficiency improvements and fixed some bugs based on wrong assumptions
Changed to simpler data structures for faster update.
Added an event handler for joints to only update points if the joints' transforms changed.
Fixed bug that assumed that the mesh and the parent joint have the same transform.
Fixed bug that assumed that the joints transforms are defined w.r.t. the scene.